golang 1.14 doesn't like calling NewFile on existing files. Port solution over from Terraform
